fix: 당첨금 저장시 오버플로우 발생가능성 제거

+++ src/main/java/lotto/domain/PurchasedLotto.java
@@ -9,7 +9,7 @@ public class PurchasedLotto {
     private List<Lotto> lottos;
     private Map<Prize, Integer> prizes;
     private int usedMoney;
-    private int earnedMoney;
+    private long earnedMoney;
 
     public PurchasedLotto() {
         this.lottos = new ArrayList<>();
